# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  model.AuthResponse:
    properties:
      expires_at:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/model.User'
    type: object
  model.Incident:
    properties:
      address:
        type: string
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      reporter_id:
        type: string
      severity:
        $ref: '#/definitions/model.IncidentSeverity'
      status:
        $ref: '#/definitions/model.IncidentStatus'
      title:
        type: string
      updated_at:
        type: string
    type: object
  model.IncidentSeverity:
    enum:
    - Low
    - Moderate
    - High
    - Critical
    type: string
    x-enum-varnames:
    - SeverityLow
    - SeverityModerate
    - SeverityHigh
    - SeverityCritical
  model.IncidentStatus:
    enum:
    - Pending
    - Verified
    - Rejected
    - Resolved
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusVerified
    - StatusRejected
    - StatusResolved
  model.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  model.ReputationLog:
    properties:
      change_amount:
        type: integer
      created_at:
        type: string
      current_score:
        type: integer
      id:
        type: string
      reason:
        type: string
      user_id:
        type: string
    type: object
  model.UpdateProfileRequest:
    properties:
      allergies:
        type: string
      avatar_url:
        type: string
      blood_type:
        type: string
      domicile:
        type: string
      emergency_hospital:
        type: string
      guardian_name:
        type: string
      guardian_phone:
        type: string
      medical_notes:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  model.User:
    properties:
      allergies:
        type: string
      avatar_url:
        type: string
      blood_type:
        type: string
      created_at:
        type: string
      domicile:
        type: string
      email:
        type: string
      emergency_hospital:
        type: string
      guardian_name:
        type: string
      guardian_phone:
        type: string
      id:
        type: string
      medical_notes:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/model.UserRole'
      status:
        description: Active, Suspended
        type: string
      trust_score:
        type: integer
      updated_at:
        type: string
    type: object
  model.UserReputationResponse:
    properties:
      benefits:
        items:
          type: string
        type: array
      description:
        type: string
      history:
        items:
          $ref: '#/definitions/model.ReputationLog'
        type: array
      sos_readiness_rate:
        type: number
      tier_color:
        type: string
      tier_label:
        type: string
      tier_status:
        type: string
      trust_score:
        type: integer
      user_id:
        type: string
      verified_reports_count:
        type: integer
      violations_count:
        type: integer
    type: object
  model.UserRole:
    enum:
    - User
    - Admin
    - Superadmin
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleAdmin
    - RoleSuperadmin
  response.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  view.IncidentResponse:
    properties:
      category:
        type: string
      description:
        type: string
      id:
        type: string
      location:
        $ref: '#/definitions/view.LocationView'
      reported_at:
        type: string
      severity:
        $ref: '#/definitions/model.IncidentSeverity'
      status:
        $ref: '#/definitions/model.IncidentStatus'
      title:
        type: string
    type: object
  view.LocationView:
    properties:
      address:
        type: string
      latitude:
        type: number
      longitude:
        type: number
    type: object
host: localhost:8080
info:
  contact:
    email: dev@jalanaman.id
    name: Tim Pengembang JalanAman
    url: https://github.com/Muhfaizr21/Jalan-Aman
  description: Dokumentasi interaktif OpenAPI/Swagger untuk API Gateway JalanAman.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/Muhfaizr21/Jalan-Aman
  title: JalanAman API Gateway
  version: 1.0.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentikasi dengan email dan password, menghasilkan token JWT dan
        cookie sesi
      parameters:
      - description: Kredensial Login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Login Pengguna & Superadmin
      tags:
      - Authentication
  /auth/logout:
    post:
      description: Menghapus cookie sesi autentikasi pengguna
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Logout Pengguna
      tags:
      - Authentication
  /auth/me:
    get:
      description: Mengambil profil pengguna/superadmin yang sedang login berdasarkan
        token JWT
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Profil Pengguna Saat Ini
      tags:
      - Authentication
  /health:
    get:
      consumes:
      - application/json
      description: Memeriksa ketersediaan API Gateway dan koneksi aktif ke database
        PostgreSQL
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Pemeriksaan Status Layanan & Database
      tags:
      - System & Health
  /incidents:
    get:
      consumes:
      - application/json
      description: Mengambil seluruh catatan insiden yang dilaporkan masyarakat
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/view.IncidentResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Daftar Seluruh Insiden
      tags:
      - Incidents
    post:
      consumes:
      - application/json
      description: Membuat dan mencatat laporan insiden jalan baru (begal, jalan rusak,
        lampu padam, kecelakaan)
      parameters:
      - description: Data Laporan Insiden
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Incident'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/view.IncidentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Laporkan Insiden Baru
      tags:
      - Incidents
  /incidents/{id}:
    get:
      consumes:
      - application/json
      description: Mengambil data spesifik insiden berdasarkan ID laporan
      parameters:
      - description: ID Insiden
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/view.IncidentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Detail Insiden Berdasarkan ID
      tags:
      - Incidents
  /users/account:
    delete:
      description: Menghapus akun pengguna secara permanen dari basis data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Hapus Akun Pengguna
      tags:
      - Authentication
  /users/profile:
    put:
      consumes:
      - application/json
      description: Memperbarui data diri, kontak, ID Medis, dan kontak pengawal darurat
      parameters:
      - description: Payload Profil
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Perbarui Profil Pengguna
      tags:
      - Authentication
  /users/reputation:
    get:
      description: Mengambil skor reputasi, tier keamanan, dan log riwayat audit dari
        PostgreSQL
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.UserReputationResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Reputasi & Tingkat Kepercayaan Pengguna
      tags:
      - Authentication
swagger: "2.0"